[Feat] Add Castbar Support to GCDBar
This is a refactoring of my main branch to try to break things into more manageable chunks.
