+++ src/UI/PluginUI.cs
@@ -1,4 +1,4 @@
-using Dalamud.Interface;
+﻿using Dalamud.Interface;
 using Dalamud.Interface.Animation;
 using Dalamud.Interface.Animation.EasingFunctions;
 using GCDTracker.Data;
@@ -6,14 +6,13 @@ using ImGuiNET;
 using System;
 using System.Numerics;
 
-namespace GCDTracker
-{
-    public class PluginUI
-    {
+namespace GCDTracker.UI {
+    public class PluginUI {
         public bool IsVisible { get; set; }
         private readonly Easing alertAnimEnabled;
         private readonly Easing alertAnimPos;
-        public GCDWheel gcd;
+        public GCDDisplay gcd;
+        public GCDHelper helper;       
         public ComboTracker ct;
         public Configuration conf;
 
@@ -54,7 +53,7 @@ namespace GCDTracker
             if (conf.WheelEnabled && !noUI && (conf.WindowMoveableGW ||
                 (enabledJobGW
                     && (conf.ShowOutOfCombat || inCombat)
-                    && (!conf.ShowOnlyGCDRunning || (gcd.idleTimerAccum < gcd.GCDTimeoutBuffer && !gcd.lastActionTP))
+                    && (!conf.ShowOnlyGCDRunning || (helper.idleTimerAccum < helper.GCDTimeoutBuffer && !helper.lastActionTP))
                     ))) {
                 SetupWindow("GCDTracker_GCDWheel", conf.WindowMoveableGW);
                 gcd.DrawGCDWheel(this);
@@ -64,7 +63,7 @@ namespace GCDTracker
             if (conf.BarEnabled && !noUI && (conf.BarWindowMoveable ||
                 (enabledJobGB
                     && (conf.ShowOutOfCombat || inCombat)
-                    && (!conf.ShowOnlyGCDRunning || (gcd.idleTimerAccum < gcd.GCDTimeoutBuffer && !gcd.lastActionTP))
+                    && (!conf.ShowOnlyGCDRunning || (helper.idleTimerAccum < helper.GCDTimeoutBuffer && !helper.lastActionTP))
                     ))) {
                 SetupWindow("GCDTracker_Bar", conf.BarWindowMoveable);
                 
@@ -337,30 +336,6 @@ namespace GCDTracker
             draw.Flags &= ~ImDrawListFlags.AntiAliasedFill;
             draw.AddRect(start, end, ImGui.GetColorU32(color), 0, ImDrawFlags.None, thickness);
             draw.Flags = originalFlags;
-        }
-        public void DrawDebugText(float relx, float rely, float textSize, Vector4 textCol, Vector4 backCol, string debugText) {
-            ImGui.PushFont(UiBuilder.MonoFont);
-            ImGui.SetWindowFontScale(textSize);
-
-            var textSz = ImGui.CalcTextSize(debugText);
-            var textStartPos =
-                w_cent
-                - (w_size / 2)
-                + new Vector2(w_size.X * relx, w_size.Y * rely)
-                - (textSz / 2);
-            var padding = new Vector2(10, 5) * textSize;
-
-            draw.AddRectFilled(
-                textStartPos - padding,
-                textStartPos + textSz + padding,
-                ImGui.GetColorU32(backCol), 10f);
-            draw.AddText(
-                textStartPos,
-                ImGui.GetColorU32(textCol),
-                debugText);
-
-            ImGui.SetWindowFontScale(1f);
-            ImGui.PopFont();
         }
     }
 }
